requirements.txt
fastapi>=0.95.0,<0.105.0
uvicorn>=0.22.0,<0.24.0
streamlit>=1.24.0,<1.29.0
pandas>=2.0.0,<2.2.0
langchain>=0.0.310,<0.0.320
langchain-openai
openai
python-dotenv>=1.0.0,<1.1.0
pydantic<2.0.0
plotly